backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.4
pdfminer.six==20231228
pypdf==4.3.1
python-docx==1.1.2
pydantic==2.8.2
gunicorn==21.2.0
openai==1.51.0
anthropic==0.36.0
httpx==0.27.2
tenacity==8.4.2
